Changed buttons, was able to pull id from button in javascript, issue with the get request.

+++ src/main/java/com/example/service/QuestionService.java
@@ -51,7 +51,7 @@ public class QuestionService {
         return testQuestions;
     }
 
-    public Question getQuestion(Long questionId) {
+    public Question getQuestion(int questionId) {
         Question question = questionRepository.findById(questionId);
         return question;
     }
